module-1/15_Review_Day/exercise/src/test/java/com/techelevator/crm/PetTests.java
package com.techelevator.crm;

import org.junit.Assert;
import org.junit.Test;
import java.util.ArrayList;
import java.util.List;

public class PetTests {


    @Test
    public void listVaccinations() {
        //arrange
        Pet classTest = new Pet("Test Name", "Test Species");
        //act
        List<String> testList = new ArrayList<>();
        testList.add("Rabies");
        testList.add("Distemper");
        testList.add("Parvo");
        classTest.setVaccinations(testList);
        //assert
        Assert.assertEquals("Rabies, Distemper, Parvo", classTest.listVaccinations());
    }
}
